refactor(structure): move scales into scales folder
