+++ src/scales/quantile.ts
