+++ src/tick-methods/d3-linear.ts
@@ -1,8 +1,8 @@
 import { head, isNil, last } from '@antv/util';
 import { ScaleConfig } from '../types';
-import { d3Linear as d3LinearMethod } from '../util/d3-linear';
-import { interval } from '../util/interval';
-import { strictLimit } from '../util/strict-limit';
+import { d3Linear as d3LinearMethod } from '../utils/d3-linear';
+import { interval } from '../utils/interval';
+import { strictLimit } from '../utils/strict-limit';
 
 export function d3Linear(cfg: ScaleConfig): number[] {
   const { min, max, tickInterval, minLimit, maxLimit } = cfg;
