+++ src/tick-methods/index.ts
