+++ src/tick-methods/linear.ts
@@ -1,8 +1,8 @@
 import { head, isNil, last } from '@antv/util';
 import { ScaleConfig } from '../types';
-import { extended } from '../util/extended';
-import { interval } from '../util/interval';
-import { strictLimit } from '../util/strict-limit';
+import { extended } from '../utils/extended';
+import { interval } from '../utils/interval';
+import { strictLimit } from '../utils/strict-limit';
 
 /**
  * 计算线性的 ticks，使用 wilkinson extended 方法
