+++ src/utils/strict-limit.ts
